+++ client/src/pages/landingpage/LandingPage.tsx
@@ -2,7 +2,6 @@ import { Link } from 'react-router-dom';
 
 import PortfolioSection from '@/components/home-section/portfolio-section/PortfolioSection';
 import CommunitySection from '@/components/home-section/community-section/CommunitySection';
-import Header from '@/components/header/Header';
 
 import { HomeWrapper, HomeContentsWrapper } from './LandingPage.styled';
 
@@ -11,7 +10,6 @@ export default function LandingPage() {
 
   return (
     <HomeWrapper>
-      <Header />
       <HomeContentsWrapper>
         <Link to='/main'>
           <PortfolioSection />
